+++ hoover/settings/common.py
@@ -17,6 +17,8 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'django.middleware.security.SecurityMiddleware',
+    'collector.middleware.NoReferral',
+    'collector.middleware.NoCache',
 )
 
 ROOT_URLCONF = 'hoover.urls'