+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## master
 * upgrade to Django 1.10
 * "External" loader that fetches documents from a URL
+* set no-cache and no-referral response headers
 
 ## 0.1 (2016-04-28)
 * initial release